refactor: HomeSlides structure and styles
